# openapi/character-asset.openapi.json
{
  "openapi": "3.2.1",
...
  "info": {
    "title": "Character-Asset API",
    "version": "0.3.0",
    "description": "REST API for spec-driven 2.5D character generation, segmentation, rigging, 8-direction profiles, animation, equipment, preview, QC and export."
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development"
    }
  ],
...
  "paths": {
    "/projects": {
      "post": {
        "operationId": "project.create",
        "summary": "Create project",
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "project"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "project.list",
        "summary": "List projects",
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "projects"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters/specs": {
      "post": {
        "operationId": "character.create_spec",
        "summary": "Create character spec",
        "tags": [
          "Character"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "project_id": {
                    "type": "string",
                    "pattern": "^proj_[A-Za-z0-9_-]+$"
                  },
                  "character_id": {
                    "type": "string",
                    "pattern": "^char_[A-Za-z0-9_-]+$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "style_profile": {
                    "type": "string",
                    "minLength": 1
                  },
                  "proportion_profile": {
                    "type": "string",
                    "minLength": 1
                  },
                  "visual": {
                    "type": "object",
                    "properties": {
                      "gender_presentation": {
                        "type": "string"
                      },
                      "age_style": {
                        "type": "string"
                      },
                      "silhouette": {
                        "type": "string"
                      },
                      "palette": {
                        "type": "object",
                        "properties": {
                          "primary": {
                            "type": "string",
                            "pattern": "^#[0-9A-Fa-f]{6}$"
                          },
                          "secondary": {
                            "type": "string",
                            "pattern": "^#[0-9A-Fa-f]{6}$"
                          },
                          "accent": {
                            "type": "string",
                            "pattern": "^#[0-9A-Fa-f]{6}$"
                          }
                        },
                        "additionalProperties": true
                      },
                      "prompt_notes": {
                        "type": "string",
                        "maxLength": 4000
                      }
                    },
                    "additionalProperties": false
                  },
                  "views_required": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "S",
                        "SW",
                        "W",
                        "NW",
                        "N",
                        "NE",
                        "E",
                        "SE"
                      ]
                    },
                    "minItems": 1,
                    "uniqueItems": true
                  },
                  "equipment_slots": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "head",
                        "weapon",
                        "shield",
                        "back",
                        "fx",
                        "custom"
                      ]
                    },
                    "uniqueItems": true
                  },
                  "rig_preset": {
                    "type": "string",
                    "minLength": 1
                  },
                  "motion_preset": {
                    "type": "string",
                    "minLength": 1
                  },
                  "output_constraints": {
                    "type": "object",
                    "properties": {
                      "canvas_width": {
                        "type": "integer",
                        "minimum": 64,
                        "maximum": 4096,
                        "default": 512
                      },
                      "canvas_height": {
                        "type": "integer",
                        "minimum": 64,
                        "maximum": 4096,
                        "default": 512
                      },
                      "transparent_background": {
                        "type": "boolean",
                        "default": true
                      },
                      "padding_px": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 512,
                        "default": 16
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false,
                "required": [
                  "project_id",
                  "character_id",
                  "name",
                  "style_profile",
                  "proportion_profile",
                  "views_required",
                  "rig_preset",
                  "motion_preset"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "spec": {
                      "$ref": "#/components/schemas/CharacterSpec"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "spec"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters/{character_id}/parts/{part_id}:approve": {
      "post": {
        "operationId": "parts.approve",
        "summary": "Approve part",
        "tags": [
          "Parts"
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "approved": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "part": {
                      "$ref": "#/components/schemas/PartAsset"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "part"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters/{character_id}/rig:validate": {
      "post": {
        "operationId": "rig.validate",
        "summary": "Validate rig",
        "tags": [
          "Rig"
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "warnings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "valid",
                    "warnings",
                    "errors"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters/{character_id}/directions:validate": {
      "post": {
        "operationId": "direction.validate",
        "summary": "Validate directions",
        "tags": [
          "Directions"
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "directions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "S",
                        "SW",
                        "W",
                        "NW",
                        "N",
                        "NE",
                        "E",
                        "SE"
                      ]
                    },
                    "uniqueItems": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "warnings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "valid",
                    "warnings",
                    "errors"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "properties": {},
            "additionalProperties": true
          }
        },
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "string",
            "pattern": "^proj_[A-Za-z0-9_-]+$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "schema_version": {
            "type": "string",
            "default": "1.0.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "project_id",
          "name",
          "schema_version",
          "created_at",
          "updated_at"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": []
}